+++ ceammc/ext/src/list/list_choice.h
@@ -6,6 +6,9 @@
 using namespace ceammc;
 
 class ListChoice : public BaseObject {
+    BoolProperty* no_repeat_;
+    size_t prev_idx_;
+
 public:
     ListChoice(const PdArgs& a);
     void onList(const AtomList& l);